+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 **Changes**
 - Upgrade the Bootstrap CSS framework to version 4.
 
+**Removed**
+- Support for Elastic Search versions less than 5.
+
 ## 0.8.1 - 2017-12-10
 
 **Added**